# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/deposits/top-deposits": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deposits"
                ],
                "summary": "aggregates deposit amounts \u0026\u0026 validator counts by depositor (tx sender) and returns top N by validator counts.",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of results to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "total_deposit",
                        "description": "Sort field (total_deposit,depositor_address,validators_total, slashed, voluntary_exited, active)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (asc|desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/deposits/top-withdrawals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Deposits"
                ],
                "summary": "aggregates deposit amounts \u0026\u0026 validator counts by withdrawal address and returns top N by validator counts.",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Number of results to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "total_deposit",
                        "description": "Sort field (total_deposit,withdrawal_address,validators_total, slashed, voluntary_exited, active)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (asc|desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rewards": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rewards"
                ],
                "summary": "Get total rewards (EL+CL) for validators from Today's rewards from UTC 0:00 to the present.",
                "parameters": [
                    {
                        "description": "Validators request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.RewardsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rewards/by-address": {
            "post": {
                "description": "Looks up validators funded by withdrawal or deposit address and returns the summed rewards for those validators.//",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rewards"
                ],
                "summary": "Get aggregated validator rewards (EL+CL) per withdrawal or deposit address.",
                "parameters": [
                    {
                        "description": "Addresses request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.AddressRewardsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rewards/network": {
            "get": {
                "description": "Uses cached consensus/execution rewards to calculate global CL/EL totals and a daily APR estimate.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rewards"
                ],
                "summary": "Get total validator rewards for the latest 24h window",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.AddressRewardsRequest": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
        "server.RewardsRequest": {
            "type": "object",
            "required": [
                "validators"
            ],
            "properties": {
                "validators": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}